+++ src/API/GoogleSearchAPI.java
@@ -8,7 +8,6 @@ import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
 import org.jsoup.select.Elements;
 
-
 public class GoogleSearchAPI {
     
     public int findNumberOfAlbums(String artist) {